deploy_tiktalik/requirements.txt
pytest
tiktalik
python-dotenv